Update: Redis Store Token

+++ auth-service/src/main/java/com/tienhuynh/auth_service/model/LoginResponse.java
@@ -10,6 +10,6 @@ public class LoginResponse {
     private String accessToken;
     private String refreshToken;
     private String tokenType;
-    private int expiresIn;
+    private long expiresIn;
     private String msg;
 }
